refactor: solana rpc apis
