+++ packages/web3-providers/package.json
@@ -21,6 +21,13 @@
     "@masknet/web3-shared-flow": "workspace:*",
     "@masknet/web3-shared-solana": "workspace:*",
     "@masknet/icons": "workspace:*",
+    "@metaplex-foundation/mpl-token-metadata": "^1.1.0",
+    "@metaplex/js": "^4.11.7",
+    "@solana/buffer-layout": "^4.0.0",
+    "@solana/spl-name-service": "^0.1.3",
+    "@solana/spl-token": "^0.1.8",
+    "@solana/spl-token-registry": "^0.2.1301",
+    "@solana/web3.js": "^1.44.2",
     "@types/socket.io-client": "^1.4.36",
     "bignumber.js": "9.0.2",
     "date-fns": "^2.28.0",
