+++ packages/web3-providers/src/solana/index.ts
@@ -1,0 +1,2 @@
+export * from './apis/FungibleAPI'
+export * from './apis/NonFungibleAPI'
